refactor: modularize interpolation and state helpers
