+++ src/hooks/interpolation/math.ts
@@ -1,0 +1,6 @@
+export function kToAlpha(k: number | undefined, dt: number): number {
+  if (!k || k <= 0 || dt <= 0) {
+    return 0;
+  }
+  return 1 - Math.exp(-k * dt);
+}
